tideguard-backend/requirements.txt
Flask==2.3.3
Flask-SocketIO==5.3.6
Flask-CORS==4.0.0
python-dotenv==1.0.0
twilio==8.10.0
python-socketio==5.9.0
eventlet==0.33.3
